requirements.txt
python-dotenv
feedparser
google_auth_oauthlib
google-api-python-client